requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.5
django-filter==22.1
django-rest-framework==0.1.0
django-rest-swagger==2.2.0
django-tracking2==0.5.1
djangorestframework==3.14.0
djangorestframework-bulk==0.2.1
drf-writable-nested==0.7.0
et-xmlfile==1.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
openapi-codec==1.3.2
openpyxl==3.1.0
Pillow==9.4.0
pytz==2022.7.1
random-username==1.0.2
requests==2.28.2
simplejson==3.18.3
six==1.16.0
sqlparse==0.4.3
uritemplate==4.1.1
urllib3==1.26.14
